(#4) Create a Webhoseio module and move the main class into the module, renamed as Client.
